sahar/popup-inner-styling (#87)
* Update CSS and JSX of Experience Form

Styles and layout of CreateExperienceForm have been updated for better user interface and experience. Changes include modifying the checkbox for "Present experience" input and enhancing button styling. Additionally, overall clean-up of the code spacing has been done to improve code readability.

* Add background styling to content in index.css

Appended width, background-image, background-size, and background-repeat properties to the .content class in index.css. This change was made to personalize the design and improve the visual appearance of the content section.