sahar/account-page-card-layout (#108)
* Add responsive design to Profile and Experience pages

The CSS for the ProfilePage and ExperienceCard components has been updated. The minimum width has been set for better responsiveness. Additionally, media queries have been added to handle different screen sizes, improving the mobile and tablet experience of the pages.

* Add styling to "Update Password" button

The "Update Password" button in UserPage.jsx is now given a unique class name and CSS rules associated with it. This change enhances the user interface by adding proper formatting and color schemes, improving the overall design and user experience of the button for readability and functionality.